requirements.txt
kaggle
torch
pytest
black
torchvision
matplotlib
tqdm
seaborn
rich
tensorboard
natsort
torchsummary
scikit-learn
monai
torchattacks
plotly
dash
dash-slicer
tensorflow
wandb
pytorchcv
SciencePlots
Tshrag-MetricDB
pyyaml
scipy
brainscore-vision